requirements.txt
asyncpg==0.22.0
aws-psycopg2==1.2.1
databases==0.4.1
fastapi==0.63.0
mangum==0.10.0
pydantic==1.8.1
SQLAlchemy==1.3.23
starlette==0.13.6
typing-extensions==3.7.4.3
PyYAML==5.4.1
